refactor(ports): split selector.py into one-Protocol-per-file modules
The monolithic ports/selector.py contained three Protocols
(CandidateSelectorProtocol, EvaluationPolicyProtocol,
ComponentSelectorProtocol) violating the project's one-Protocol-per-file
convention. This splits them into dedicated modules and aligns tooling.

- Extract CandidateSelectorProtocol to ports/candidate_selector.py
- Extract EvaluationPolicyProtocol to ports/evaluation_policy.py
- Extract ComponentSelectorProtocol to ports/component_selector.py
- Delete ports/selector.py; update all import consumers (11 files)
- Add import path equivalence tests to all 3 contract test files
- Fix ruff isort known-first-party from agent_workflow_suite to gepa_adk
- Add [tool.docvet] fail-on config to pyproject.toml
- Create src/gepa_adk/py.typed PEP 561 marker
- Align ty config: add include to [tool.ty.src], simplify pre-commit entry
