+++ .pre-commit-config.yaml
@@ -30,7 +30,7 @@ repos:
     hooks:
       - id: ty-check
         name: ty type check
-        entry: uv run ty check src tests
+        entry: uv run ty check
         language: system
         types: [python]
         pass_filenames: false
