+++ src/gepa_adk/adapters/candidate_selector.py
@@ -19,7 +19,7 @@ import random
 
 from gepa_adk.domain.exceptions import ConfigurationError, NoCandidateAvailableError
 from gepa_adk.domain.state import ParetoState
-from gepa_adk.ports.selector import CandidateSelectorProtocol
+from gepa_adk.ports.candidate_selector import CandidateSelectorProtocol
 
 
 class ParetoCandidateSelector:
