+++ src/gepa_adk/ports/evaluation_policy.py
@@ -1,8 +1,7 @@
-"""Protocol definition for candidate selection strategies.
+"""Protocol definition for valset evaluation strategies.
 
-Note:
-    This module defines the async contract for candidate selection strategies
-    and evaluation policies.
+Attributes:
+    EvaluationPolicyProtocol: Protocol for valset evaluation strategies.
 """
 
 from __future__ import annotations
@@ -10,46 +9,6 @@ from __future__ import annotations
 from typing import Protocol, Sequence, runtime_checkable
 
 from gepa_adk.domain.state import ParetoState
-
-
-@runtime_checkable
-class CandidateSelectorProtocol(Protocol):
-    """Async protocol for candidate selection strategies.
-
-    Note:
-        Adapters implementing this protocol provide strategies for selecting
-        candidates from the Pareto frontier for mutation.
-
-    Examples:
-        ```python
-        class Selector:
-            async def select_candidate(self, state: ParetoState) -> int:
-                return 0
-        ```
-    """
-
-    async def select_candidate(self, state: ParetoState) -> int:
-        """Select a candidate index for mutation.
-
-        Args:
-            state: Current evolution state with Pareto frontier tracking.
-
-        Returns:
-            Index of selected candidate.
-
-        Raises:
-            NoCandidateAvailableError: If state has no candidates.
-
-        Note:
-            Outputs a candidate index from the frontier for mutation,
-            enabling Pareto-aware selection strategies.
-
-        Examples:
-            ```python
-            candidate_idx = await selector.select_candidate(state)
-            ```
-        """
-        ...
 
 
 @runtime_checkable
@@ -154,49 +113,4 @@ class EvaluationPolicyProtocol(Protocol):
         ...
 
 
-@runtime_checkable
-class ComponentSelectorProtocol(Protocol):
-    """Async protocol for component selection strategies.
-
-    Note:
-        Adapters implementing this protocol determine which candidate components
-        to update during mutation, enabling flexible evolution strategies.
-
-    Examples:
-        ```python
-        class MySelector:
-            async def select_components(
-                self, components: list[str], iteration: int, candidate_idx: int
-            ) -> list[str]:
-                return components[:1]
-        ```
-    """
-
-    async def select_components(
-        self, components: list[str], iteration: int, candidate_idx: int
-    ) -> list[str]:
-        """Select components to update for the current iteration.
-
-        Args:
-            components: List of available component keys (e.g. ["instruction", "input_schema"]).
-            iteration: Current global iteration number (0-based).
-            candidate_idx: Index of the candidate being evolved.
-
-        Returns:
-            List of component keys to update.
-
-        Raises:
-            ValueError: If components list is empty.
-
-        Note:
-            Outputs a list of component keys to update, enabling selective
-            mutation of specific candidate components.
-
-        Examples:
-            ```python
-            selected = await selector.select_components(
-                components=["instruction", "schema"], iteration=1, candidate_idx=0
-            )
-            ```
-        """
-        ...
+__all__ = ["EvaluationPolicyProtocol"]
